Added CDSegment64 and CDSection64.

+++ CDSection.h
@@ -3,7 +3,7 @@
 //  This file is part of class-dump, a utility for examining the Objective-C segment of Mach-O files.
 //  Copyright (C) 1997-1998, 2000-2001, 2004-2007  Steve Nygard
 
-#import <Foundation/NSObject.h>
+#import <Foundation/Foundation.h>
 #include <mach-o/loader.h>
 
 @class NSString;
@@ -26,9 +26,9 @@
 
 - (NSString *)segmentName;
 - (NSString *)sectionName;
-- (unsigned long)addr;
-- (unsigned long)size;
-- (unsigned long)offset;
+- (uint32_t)addr;
+- (uint32_t)size;
+- (uint32_t)offset;
 
 - (const void *)dataPointer; // Has no access to the mach-o file
